Add support for min and max counts on items
This adds the final piece of definite work -- min and max counts on the items.
It also changes the returned results of a roll, compacting the results set into
a list of LootResults - a count and a name. This makes the output a lot cleaner.
Instead of listing eg

1 apple
1 banana
4 apple

It now lists
5 apple
1 banana

It decides simply by name, so some complex ideas are possible with inner tables.